fix(): delete_issue
