+++ cypress/e2e/add-questions/AddQuestion.cy.ts
@@ -11,7 +11,7 @@ describe('Test for section-details', () => {
 
   it('checks, add question button should be visible and have correct text', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -22,7 +22,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 6000 }).should(
       'have.text',
       `+ ${addQuestion.addQuestion}`
@@ -31,7 +31,7 @@ describe('Test for section-details', () => {
 
   it('checks, heading of add question should be visible and have correct text', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -42,7 +42,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 20000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -56,7 +56,7 @@ describe('Test for section-details', () => {
   })
   it('checks, redirect to section page after clicking on breadscrum link', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -67,7 +67,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 20000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -83,7 +83,7 @@ describe('Test for section-details', () => {
 
   it('Verifying MCQ to have Check Box in options', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -94,7 +94,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 20000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -130,7 +130,7 @@ describe('Test for section-details', () => {
 
   it('Verifying Single Choice to have Radio Button in options', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -141,7 +141,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 6000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -176,7 +176,7 @@ describe('Test for section-details', () => {
 
   it('Verifying Text to have Textarea in options and should be in focused after interaction', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -187,7 +187,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 6000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -225,7 +225,7 @@ describe('Test for section-details', () => {
   let lengthBefore: number
   it('Verifying if Add Option functionality Working on Options', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -236,7 +236,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 6000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -273,7 +273,7 @@ describe('Test for section-details', () => {
   })
   it('Verifying if Delete button should be visible', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -284,7 +284,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 6000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -316,7 +316,7 @@ describe('Test for section-details', () => {
   })
   it('Verifying if Delete functionality Working on Options', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -327,7 +327,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#add-question', { timeout: 6000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
       .click()
@@ -364,7 +364,7 @@ describe('Test for section-details', () => {
   })
   it('On Save and Add More visit the Add Question Page', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -375,7 +375,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
 
     cy.get('#add-question', { timeout: 6000 })
       .should('have.text', `+ ${addQuestion.addQuestion}`)
@@ -410,7 +410,7 @@ describe('Test for section-details', () => {
   })
   it('On Save and Continue visit the Sections Page', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -421,7 +421,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#section-details-heading', { timeout: 6000 })
       .should('be.visible', { timeout: 4000 })
       .should('have.text', section1)
@@ -456,7 +456,7 @@ describe('Test for section-details', () => {
   })
   it('Verifying if Question is Empty or not', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -467,7 +467,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#section-details-heading', { timeout: 6000 })
       .should('be.visible', { timeout: 4000 })
       .should('have.text', section1)
@@ -489,7 +489,7 @@ describe('Test for section-details', () => {
   })
   it('Verifying if any Option is empty or not', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -500,7 +500,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('#section-details-heading', { timeout: 6000 })
       .should('be.visible', { timeout: 4000 })
       .should('have.text', section1)
@@ -543,7 +543,7 @@ describe('Test for section-details', () => {
   })
   it('Checks if Delete Question button exists', () => {
     cy.get('#sections', { timeout: 6000 }).should('have.text', 'Tests').click()
-    cy.wait(2000)
+    
     cy.get('.sectionCard', { timeout: 6000 }).each(($el) => {
       cy.wrap($el).within((el) => {
         if (
@@ -554,7 +554,7 @@ describe('Test for section-details', () => {
       })
     })
     cy.get('.sectionName').contains(section1).click()
-    cy.wait(2000)
+    
     cy.get('.question', { timeout: 8000 })
       .contains('What is useRef() ?')
       .trigger('mouseover')
